Extract reusable filter function to its own file
